REGRESSION(r252523): Web Inspector: Styles: semicolon is wrongly positioned in multiline values https://bugs.webkit.org/show_bug.cgi?id=205034 <rdar://problem/57772846>
Reviewed by Brian Burg.

Wrap the value node in another `<span class="value-container">` so that the semicolon can be
included as part of the `display: inline-block;` when there are multiple lines in the value.

* UserInterface/Views/SpreadsheetStyleProperty.js:
(WI.SpreadsheetStyleProperty.prototype.update):

* UserInterface/Views/SpreadsheetCSSStyleDeclarationEditor.css:
(.spreadsheet-style-declaration-editor .property:not(.disabled).has-newline .value-container): Added.
(.spreadsheet-style-declaration-editor .property:not(.disabled).has-newline .value): Deleted.


Identifier: 218281@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253338 268f45cc-cd09-0410-ab3c-d52691b4dbfc