DFG and FTL expects String.prototype to not qualify for StringObjectUse. https://bugs.webkit.org/show_bug.cgi?id=205147 <rdar://problem/57748888>
Reviewed by Saam Barati.

JSTests:

* stress/ftl-expects-string-prototype-to-not-be-StringObjectUse.js: Added.

Source/JavaScriptCore:

Currently, String.prototype's JSType is StringObjectType.

However, in the compiler, there are a few places that expect that the
String.prototype value to not qualify as StringObjectUse.  These places are:
1. SpeculatedType.cpp's speculationFromClassInfo() will speculate SpecObjectOther
   for the StringPrototype object.
2. DFGFixupPhase.cpp's addCheckStructureForOriginalStringObjectUse() only emits a
   CheckStructure against globalObject->stringObjectStructure().  It does not
   check against String.prototype's structure.

To resolve this discrepancy, we can either do:
a. change String.prototype's JSType to something else.
b. fix the places in the compiler to accept String.prototype as StringObjectUse.

(a) is trivial and cheap to do.  (b) is doable but will result in less optimal
compiled code.  Since passing String.prototype as a StringObject is expected to
be a rare thing in JS code, it's not worth incurring the cost for (b).  In this
patch, we apply (a) to fix the discrepancy.

Also added a specialization case to FOR_EACH_JS_DYNAMIC_CAST_JS_TYPE_OVERLOAD
for jsDynamicCast<StringObject> for completeness.

* runtime/JSCast.h:
* runtime/JSType.cpp:
(WTF::printInternal):
* runtime/JSType.h:
* runtime/StringPrototype.h:



Identifier: 218360@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253432 268f45cc-cd09-0410-ab3c-d52691b4dbfc