[macOS] Swipe gesture snapshot stays too long if provisional load has not started yet when endSwipeGesture() is called https://bugs.webkit.org/show_bug.cgi?id=205206
Reviewed by Tim Horton.

Swipe gesture snapshot stays too long if provisional load has not started yet when endSwipeGesture() is called on macOS.
This is because the SnapshotRemovalTracker::eventOccurred() calls get ignored while the SnapshotRemovalTracker is paused
and the SnapshotRemovalTracker only gets unpaused once the provisional load has started. The idea is that we should ignore
any events from a previous navigation. However, the SwipeGestureEnd event is a UI-side event fired by the
ViewGestureController itself, so there is reason to ignore it. Since we ask the WebContent process to do the load in
willEndSwipeGesture(), it is possible that the provisional load has not started yet by the time endSwipeGesture() is
called. In such case, the SwipeGestureEnd event would get ignored and we would keep the snapshot until the timeout.

* UIProcess/ViewGestureController.cpp:
(WebKit::stopWaitingForEvent):
(WebKit::ViewGestureController::SnapshotRemovalTracker::eventOccurred):
(WebKit::ViewGestureController::endSwipeGesture):
(WebKit::ViewGestureController::SnapshotRemovalTracker::stopWaitingForEvent): Deleted.
* UIProcess/ViewGestureController.h:


Identifier: 218430@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253514 268f45cc-cd09-0410-ab3c-d52691b4dbfc