Structure should have a bloom filter of seen identifiers https://bugs.webkit.org/show_bug.cgi?id=205182
Reviewed by Yusuke Suzuki and Tadeu Zagallo.

This patch adds a bloom filter of seen identifiers to Structure. This usually allows
us to quickly determine if a Structure *has not* seen a particular property. Based
on some logging I added in JetStream2 and Speedometer2, 70% of calls to Structure::get
result in us returning invalidOffset (e.g, the property does not exist). This patch
allows that path to be even faster. This bloom filter is just modeling what goes inside
Structure's property table. For that reason, we don't need to consider things inside
the static property table. We reason about the static  property table inside JSObject's
property lookup.

This patch appears to be a 0.5% progression on Speedometer2.

* runtime/Structure.cpp:
(JSC::Structure::Structure):
* runtime/Structure.h:
* runtime/StructureInlines.h:
(JSC::Structure::get):
(JSC::Structure::add):


Identifier: 218433@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253517 268f45cc-cd09-0410-ab3c-d52691b4dbfc