-requestDocumentContext always returns 1 text unit more granularity than requested https://bugs.webkit.org/show_bug.cgi?id=205142 <rdar://problem/57858236>
Reviewed by Darin Adler and Wenson Hsieh.

Source/WebCore:

Fix up the code to actually determine if the specified position is at a sentence
boundary. Currently the code will always return false when asking whether the
specified position is at a sentence boundary (i.e. "end of the sentence") because
it compares it to the position of the end of the *next* sentence or the beginning
of the current sentence when selecting forward or backwards, respectively.

* editing/VisibleUnits.cpp:
(WebCore::atBoundaryOfGranularity):

Source/WebKit:

Use WebCore::atBoundaryOfGranularity() to identify each boundary so that we return the position
exactly granularityCount text units advanced from the specified position. When using sentence
granularity we do not need to round the resulting position to the nearest word because it already
falls before the next word (if there is one). For all other granularities we do what we do now
and round to the nearest word, which may cross that granularity's boundary.

Additionally, added assertions to ensure that we are passed a non-zero granularity count and a non-
null initial position. The function takes advantage of these assumptions to 1) ensure correct results
and 2) make use of a do-while loop.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::moveByGranularityRespectingWordBoundary):

Tools:

Add tests for requesting text by character, sentence, paragraph, and line granularities.
Also organized tests, demarcating sections of related tests, and renaming the existing
word granularity tests for consistency.

* TestWebKitAPI/Tests/WebKitCocoa/DocumentEditingContext.mm:
(TEST):

Identifier: 218477@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253561 268f45cc-cd09-0410-ab3c-d52691b4dbfc