WebGLRenderingContext.texImage2D() should respect EXIF orientation https://bugs.webkit.org/show_bug.cgi?id=205141
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2019-12-16
Reviewed by Simon Fraser.

Source/WebCore:

If image orientation is not the default, WebGLRenderingContext.texImage2D()
needs to draw this image into an ImageBuffer, makes a temporary Image
from the ImageBuffer then draw this temporary Image to the WebGL texture.

Test: fast/images/exif-orientation-webgl-texture.html

* html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::texSubImage2D):
(WebCore::WebGLRenderingContextBase::texImage2D):
* platform/graphics/BitmapImage.h:
* platform/graphics/Image.h:
(WebCore::Image::orientation const):

LayoutTests:

The test page uses images with different EXIF orientation. The expected
page uses a single image with no EXIF orientation then it transforms the
<canvas> elements such that it matches the image in the test page.

WebGLRenderingContext uses a trick when drawing an SVG image or images
with EXIF orientation to a WebGL texture. It draws the Image to an Image-
Buffer, creates another Image out of the ImageBuffer and then draws the
other Image to the WebGL texture.

But there can be small glitches between drawing an Image directly versus
doing the ImageBuffer trick. So the expected page will use an SVG image
to ensure the same code path is used for both the test and the expected
pages.

This SVG image includes the jpeg image with no EXIF orientation but as a
data uri. Also the script has to wait after loading the SVG image till
the bitmap image is loaded from the data uri encoded data.

* fast/images/exif-orientation-webgl-texture-expected.html: Added.
* fast/images/exif-orientation-webgl-texture.html: Added.
* fast/images/resources/webgl-draw-image.js: Added.
* platform/win/TestExpectations:
All webgl tests are skipped on Windows.

Identifier: 218517@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253610 268f45cc-cd09-0410-ab3c-d52691b4dbfc