REGRESSION: animations/leak-document-with-css-animation.html is very flaky https://bugs.webkit.org/show_bug.cgi?id=205299
While this test has proven to be solid running in isolation, it appears that the difference of live documents
before and after the iframe has been destroyed may be more than one. In fact, in reported failures, there always
seems to be 2 live documents fewer after than before, instead of the expected single document.

Since internals.numberOfLiveDocuments() returns the number of live documents throughout the life of the
Web process, we now force garbage collection every time we query internals.numberOfLiveDocuments(). This
should ensure that documents pending destruction due to garbage collection to be destroyed before we take
a snapshort of the number of live documents.

Patch by Antoine Quint <graouts@apple.com> on 2019-12-18
Reviewed by Antti Koivisto.

* TestExpectations:
* animations/leak-document-with-css-animation.html:

Identifier: 218583@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253679 268f45cc-cd09-0410-ab3c-d52691b4dbfc