WebDriver: several w3c collections tests added in r230953 are failing https://bugs.webkit.org/show_bug.cgi?id=184966
Reviewed by Brian Burg.

Source/WebKit:

This is because JSON.stringify() is not enough, we need to first clone the value using the internal JSON clone
algorithm explained in the spec. See https://w3c.github.io/webdriver/#dfn-internal-json-clone-algorithm.

Fixes: imported/w3c/webdriver/tests/execute_script/collections.py::test_arguments
       imported/w3c/webdriver/tests/execute_script/collections.py::test_file_list
       imported/w3c/webdriver/tests/execute_script/collections.py::test_html_all_collection
       imported/w3c/webdriver/tests/execute_async_script/collections.py::test_arguments
       imported/w3c/webdriver/tests/execute_async_script/collections.py::test_file_list
       imported/w3c/webdriver/tests/execute_async_script/collections.py::test_html_all_collection

* WebProcess/Automation/WebAutomationSessionProxy.js:
(WebKitAutomation.AutomationSessionProxy.prototype._jsonStringify): Call JSON.stringify(), but without the
replacer and passing the result of _jsonClone().
(WebKitAutomation.AutomationSessionProxy.prototype._isCollection): Helper to check if the value is a collection.
(WebKitAutomation.AutomationSessionProxy.prototype._checkCyclic): Throw an error if the value is a cyclic structure.
(WebKitAutomation.AutomationSessionProxy.prototype._jsonClone): Implement the internal JSON clone algorithm.
(WebKitAutomation.AutomationSessionProxy.prototype._createNodeHandle): Trhow an error if the node is stale.

WebDriverTests:

Remove expectations for tests that are now passing.

* TestExpectations.json:

Identifier: 218988@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254120 268f45cc-cd09-0410-ab3c-d52691b4dbfc